requirements.txt
Django==3.0.5
Werkzeug==1.0.1
django-environ==0.4.5
serverless-psycopg2==0.0.2
djangorestframework==3.11.0
drf-yasg==1.17.1
requests==2.23.0
